Refactor code

+++ src/main/java/com/example/javatextanalyzer/AnalyzeController.java
@@ -4,8 +4,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.TreeMap;
 
 @RestController
@@ -13,17 +11,17 @@ public class AnalyzeController {
 
     @PostMapping("/analyze")
     public Result result(@RequestBody Request request) {
-        return new Result(getTextLength(request.getText()),getWordCount(request.getText()),getCharacterCount(request.getText()));
+        return new Result(getTextLength(request.getText()), getWordCount(request.getText()), getCharacterCount(request.getText()));
     }
 
     public TextLength getTextLength(String text) {
-        long withSpaces = text.length();
-        long withoutSpaces = text.replaceAll(" ","").length();
-        return new TextLength(withSpaces,withoutSpaces);
+        int withSpaces = text.length();
+        int withoutSpaces = text.replaceAll(" ", "").length();
+        return new TextLength(withSpaces, withoutSpaces);
     }
 
-    public long getWordCount(String text) {
-        if(text == null || text.isEmpty()) {
+    public int getWordCount(String text) {
+        if (text == null || text.isEmpty()) {
             return 0;
         }
 
@@ -34,23 +32,19 @@ public class AnalyzeController {
 
     public Object[] getCharacterCount(String text) {
 
-        String cleanString = text.toLowerCase().replaceAll("[^a-z]","");
+        String cleanString = text.toLowerCase().replaceAll("[^a-z]", "");
 
         TreeMap<Character, Integer> charCountMap = new TreeMap<Character, Integer>();
 
         char[] strArray = cleanString.toCharArray();
 
-        for(char c : strArray) {
-            if(charCountMap.containsKey(c)) {
-                charCountMap.put(c, charCountMap.get(c) + 1 );
-            }else {
+        for (char c : strArray) {
+            if (charCountMap.containsKey(c)) {
+                charCountMap.put(c, charCountMap.get(c) + 1);
+            } else {
                 charCountMap.put(c, 1);
             }
         }
-
-
-
-
         return charCountMap.entrySet().toArray();
     }
